feat: implement GL shaders system
The shaders are located to the `shaders` directory. In a similar manner
to webdev (eww), the shaders are preprocessed via a preprocessor (same
as a bundler in webdev). This preprocessor is a simple python script
that supports things like `#include` and `#pragma once`. For simplicity,
the preprocessor will only support a small subset of what a full-fledged
preprocessor should do (and be prone to stuff like `# include <...>`).

Shader source files are located in `shaders/src`, and generated files
should be located at `shaders/out`. The application will watch for
changes in the output directory and hot-reload the updated shader.
The shader programmer can easily use an external program to preprocess
updated shader source files on file changes like how it is done in
webdev.